Enum config::ConfigError

pub enum ConfigError {
    Frozen,
    NotFound(String),
    PathParse(ErrorKind),
    FileParse {
        uri: Option<String>,
        cause: Box<Error + Send + Sync>,
    },
    Type {
        origin: Option<String>,
        unexpected: Unexpected,
        expected: &'static str,
        key: Option<String>,
    },
    Message(String),
    Foreign(Box<Error + Send + Sync>),
}

Represents all possible errors that can occur when working with configuration.

Variants
Frozen

Configuration is frozen and no further mutations can be made.

NotFound(String)

Configuration property was not found

PathParse(ErrorKind)

Configuration path could not be parsed.

FileParse

Configuration could not be parsed from file.

Fields of FileParse

uri: Option<String>

The URI used to access the file (if not loaded from a string). Example: /path/to/config.json

cause: Box<Error + Send + Sync>

The captured error from attempting to parse the file in its desired format. This is the actual error object from the library used for the parsing.

Type

Value could not be converted into the requested type.

Fields of Type

origin: Option<String>

The URI that references the source that the value came from. Example: /path/to/config.json or Environment or etcd://localhost

unexpected: Unexpected

What we found when parsing the value

expected: &'static str

What was expected when parsing the value

key: Option<String>

The key in the configuration hash of this value (if available where the error is generated).

Message(String)

Custom message

Foreign(Box<Error + Send + Sync>)

Unadorned error from a foreign origin.