Enum config::FileFormat

pub enum FileFormat {
    Toml,
    Json,
    Yaml,
    Ini,
    Ron,
    Json5,
}

File formats provided by the library.

Although it is possible to define custom formats using Format trait it is recommended to use FileFormat if possible.

Variants

Toml

TOML (parsed with toml)

Json

JSON (parsed with serde_json)

Yaml

YAML (parsed with yaml_rust)

Ini

INI (parsed with rust_ini)

Ron

RON (parsed with ron)

Json5

JSON5 (parsed with json5)