Enum config::ValueKind

pub enum ValueKind {
    Nil,
    Boolean(bool),
    I64(i64),
    I128(i128),
    U64(u64),
    U128(u128),
    Float(f64),
    String(String),
    Table(Map<String, Value>),
    Array(Vec<Value>),
}

Underlying kind of the configuration value.

Standard operations on a Value by users of this crate do not require knowledge of ValueKind. Introspection of underlying kind is only required when the configuration values are unstructured or do not have known types.

Variants

Nil

Boolean(bool)

I64(i64)

I128(i128)

U64(u64)

U128(u128)

Float(f64)

String(String)

Table(Map<String, Value>)

Array(Vec<Value>)