config

Function from_lua_value

pub fn from_lua_value<R>(lua: &Lua, value: Value) -> Result<R>
where
    R: DeserializeOwned,

Convert from a lua value to a deserializable type, with a slightly more helpful error message in case of failure. NOTE: the “, while processing” portion of the error messages generated here is coupled with a regex in typing.lua!