Struct config::File

pub struct File<T> where
    T: FileSource,  { /* fields omitted */ }

Methods

impl File<FileSourceString>

pub fn from_str(s: &str, format: FileFormat) -> Self

impl File<FileSourceFile>

pub fn new(name: &str, format: FileFormat) -> Self

pub fn with_name(name: &str) -> Self

Given the basename of a file, will attempt to locate a file by setting its extension to a registered format.

impl<T: FileSource> File<T>

pub fn format(self, format: FileFormat) -> Self

pub fn required(self, required: bool) -> Self