Struct config::Value

pub struct Value {
    pub kind: ValueKind,
    /* private fields */
}

A configuration value.

Fields
kind: ValueKind

Underlying kind of the configuration value.

Implementations

impl Value

pub fn new<V>(origin: Option<&String>, kind: V) -> Self
where V: Into<ValueKind>,

Create a new value instance that will remember its source uri.

pub fn origin(&self) -> Option<&str>

Get the description of the original location of the value.

pub fn try_deserialize<'de, T: Deserialize<'de>>(self) -> Result<T, ConfigError>

Attempt to deserialize this value into the requested type.

pub fn into_bool(self) -> Result<bool, ConfigError>

Returns self as a bool, if possible.

pub fn into_int(self) -> Result<i64, ConfigError>

Returns self into an i64, if possible.

pub fn into_int128(self) -> Result<i128, ConfigError>

Returns self into an i128, if possible.

pub fn into_uint(self) -> Result<u64, ConfigError>

Returns self into an u64, if possible.

pub fn into_uint128(self) -> Result<u128, ConfigError>

Returns self into an u128, if possible.

pub fn into_float(self) -> Result<f64, ConfigError>

Returns self into a f64, if possible.

pub fn into_string(self) -> Result<String, ConfigError>

Returns self into a string, if possible.

pub fn into_array(self) -> Result<Vec<Self>, ConfigError>

Returns self into an array, if possible

pub fn into_table(self) -> Result<Map<String, Self>, ConfigError>

If the Value is a Table, returns the associated Map.