Module config::types

Internal types used to represent a configuration and corresponding primitives to browse it

Structs

Config
The top-level Config type that represents a configuration

Setting
A Setting representation. Settings have a name and a value.

Enums

ScalarValue
The scalar values representation. Scalar values bind directly to Rust primitive types.

Value
A type representing a generic value. Settings store Values.

Type Definitions

ArrayValue
The type used to represent the scalars inside an array. An array can only store scalar values of the same type.

ListValue
The type used to represent the generic values inside a list. Lists are heterogeneous and can store any type of value, including other lists.

SettingsList
Settings list representation. Associates settings to their names.