Crate config

Structs

CallbackSignature

CallbackSignature is a bit sugar to aid with statically typing event callback function invocation.

LuaConfig

Statics

VALIDATE_ONLY

VALIDATION_FAILED

Functions

any_err

Helper for mapping back to lua errors

decorate_callback_name

does_callback_allow_multiple

from_lua_value

Convert from a lua value to a deserializable type, with a slightly more helpful error message in case of failure.

get_or_create_module

get_or_create_sub_module

Given a name path like foo or foo.bar.baz, sets up the module registry hierarchy to instantiate that path. Returns the leaf node of that path to allow the caller to register/assign functions etc. into it

is_validating

load_config

register

serialize_options

set_max_age

set_max_spare

set_max_use

set_policy_path

set_validation_failed

validation_failed

Type Aliases

RegisterFunc