Crate config

Modules

utils

Structs

Authority

BlockSynchronizerParameters

Committee

ConsensusAPIGrpcParameters

Parameters

Holds all the node properties. An example is provided to showcase the usage and deserialization from a json file. To define a Duration on the property file can use either milliseconds or seconds (e.x 5s, 10ms , 2000ms).

PrimaryAddresses

PrometheusMetricsParameters

WorkerCache

WorkerIndex

WorkerInfo

Enums

CommitteeUpdateError

ConfigError

Traits

Export

Import

Type Definitions

Epoch

The epoch number.

SharedCommittee

SharedWorkerCache

Stake

WorkerId