Struct config::Config

pub struct Config {
    pub cache: Value,
    // some fields omitted
}

A prioritized configuration repository. It maintains a set of configuration sources, fetches values to populate those, and provides them according to the source's priority.

Fields
cache: Value

Root of the cached configuration.

Methods

impl Config

pub fn new() -> Self

pub fn merge<T>(&mut self, source: T) -> Result<&mut Config, ConfigError> where
T: 'static,
T: Source + Send + Sync,

Merge in a configuration property source.

pub fn refresh(&mut self) -> Result<&mut Config, ConfigError>

Refresh the configuration cache with fresh data from added sources.

Configuration is automatically refreshed after a mutation operation (set, merge, set_default, etc.).

pub fn set_default<T>(
&mut self,
key: &str,
value: T
) -> Result<&mut Config, ConfigError> where
T: Into<Value>,

pub fn set<T>(
&mut self,
key: &str,
value: T
) -> Result<&mut Config, ConfigError> where
T: Into<Value>,

pub fn get<'de, T: Deserialize<'de>>(
&self,
key: &'de str
) -> Result<T, ConfigError>

pub fn get_str(&self, key: &str) -> Result<String, ConfigError>

pub fn get_int(&self, key: &str) -> Result<i64, ConfigError>

pub fn get_float(&self, key: &str) -> Result<f64, ConfigError>

pub fn get_bool(&self, key: &str) -> Result<bool, ConfigError>

pub fn get_table(
&self,
key: &str
) -> Result<HashMap<String, Value>, ConfigError>

pub fn get_array(&self, key: &str) -> Result<Vec<Value>, ConfigError>

pub fn try_into<'de, T: Deserialize<'de>>(self) -> Result<T, ConfigError>

Attempt to deserialize the entire configuration into the requested type.

pub fn deserialize<'de, T: Deserialize<'de>>(self) -> Result<T, ConfigError>

Deprecated since 0.7.0
: please use 'try_into' instead