Struct config::Config

pub struct Config {
    pub cache: Value,
    /* private fields */
}

A prioritized configuration repository. It maintains a set of configuration sources, fetches values to populate those, and provides them according to the source’s priority.

Fields
cache: Value

Root of the cached configuration.

Implementations

impl Config

pub fn builder() -> ConfigBuilder<DefaultState>

Creates new ConfigBuilder instance

pub fn merge<T>(&mut self, source: T) -> Result<&mut Self, ConfigError>
where T: 'static + Source + Send + Sync,

👎Deprecated since 0.12.0: please use ‘ConfigBuilder’ instead

Merge in a configuration property source.

pub fn with_merged<T>(self, source: T) -> Result<Self, ConfigError>
where T: 'static + Source + Send + Sync,

👎Deprecated since 0.12.0: please use ‘ConfigBuilder’ instead

Merge in a configuration property source.

pub fn refresh(&mut self) -> Result<&mut Self, ConfigError>

👎Deprecated since 0.12.0: please use ‘ConfigBuilder’ instead

Refresh the configuration cache with fresh data from added sources.

Configuration is automatically refreshed after a mutation operation (set, merge, set_default, etc.).

pub fn set_default<T>( &mut self, key: &str, value: T ) -> Result<&mut Self, ConfigError>
where T: Into<Value>,

👎Deprecated since 0.12.0: please use ‘ConfigBuilder’ instead

Set a default value at key

pub fn set<T>(&mut self, key: &str, value: T) -> Result<&mut Self, ConfigError>
where T: Into<Value>,

👎Deprecated since 0.12.0: please use ‘ConfigBuilder’ instead

Set an overwrite

This function sets an overwrite value. The overwrite value is written to the key location on every refresh()

Warning

Errors if config is frozen

pub fn set_once(&mut self, key: &str, value: Value) -> Result<(), ConfigError>

👎Deprecated since 0.12.0: please use ‘ConfigBuilder’ instead

pub fn get<'de, T: Deserialize<'de>>(&self, key: &str) -> Result<T, ConfigError>

pub fn get_string(&self, key: &str) -> Result<String, ConfigError>

pub fn get_int(&self, key: &str) -> Result<i64, ConfigError>

pub fn get_float(&self, key: &str) -> Result<f64, ConfigError>

pub fn get_bool(&self, key: &str) -> Result<bool, ConfigError>

pub fn get_table(&self, key: &str) -> Result<Map<String, Value>, ConfigError>

pub fn get_array(&self, key: &str) -> Result<Vec<Value>, ConfigError>

pub fn try_deserialize<'de, T: Deserialize<'de>>(self) -> Result<T, ConfigError>

Attempt to deserialize the entire configuration into the requested type.

pub fn try_from<T: Serialize>(from: &T) -> Result<Self, ConfigError>

Attempt to serialize the entire configuration from the given type.

pub fn deserialize<'de, T: Deserialize<'de>>(self) -> Result<T, ConfigError>

👎Deprecated since 0.7.0: please use ‘try_deserialize’ instead