Struct config::Value

pub struct Value {
    pub kind: ValueKind,
    // some fields omitted
}

A configuration value.

Fields
kind: ValueKind

Underlying kind of the configuration value.

Methods

impl Value

pub fn new<V>(origin: Option<&String>, kind: V) -> Self where
V: Into<ValueKind>,

Create a new value instance that will remember its source uri.

pub fn try_into<'de, T: Deserialize<'de>>(self) -> Result<T, ConfigError>

Attempt to deserialize this value into the requested type.

pub fn into_bool(self) -> Result<bool, ConfigError>

Returns self as a bool, if possible.

pub fn into_int(self) -> Result<i64, ConfigError>

Returns self into an i64, if possible.

pub fn into_float(self) -> Result<f64, ConfigError>

Returns self into a f64, if possible.

pub fn into_str(self) -> Result<String, ConfigError>

Returns self into a str, if possible.

pub fn into_array(self) -> Result<Vec<Value>, ConfigError>

Returns self into an array, if possible

pub fn into_table(self) -> Result<HashMap<String, Value>, ConfigError>

If the Value is a Table, returns the associated Map.